requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
arxiv==1.4.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==2.1.1
discord.py==2.2.3
exceptiongroup==1.1.1
feedparser==6.0.10
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
notion-client==2.0.0
requests==2.28.1
rsrch==0.2.3
sgmllib3k==1.0.0
sniffio==1.3.0
tqdm==4.64.1
urllib3==1.26.16
yarl==1.9.2
